ccp: support s3 to http acceleration via presigned url and batch downloading
By default, ccp will download http and https files larger than 128MiB in parallel
by relying on temp files. This can be disabled by using ccp -slow. The partition
size can be changed from 128MiB with ccp -partsize and the temp folder can
be changed with ccp -tmp or by setting TEMPDIR

The algorithm that accelerates http downloads preserves the streaming flow
of ccp, and should not require any modifications in scripts for proper use.

Another optimization for s3 is to drop https in favor for http. If applicable, all
s3 file inputs are presigned and stripped to their http equivalent URLs. This
too can be disabled with ccp -secure