ccp: cat and append-only mode flags
If ccp is writing to standard output, all the files in the source list
are concatenated to that file. This mode can also be forced for
regular files with the -cat flag.

The append flag enables append-only mode for all destination
files in the OS filesystem and currently nowhere else.

All copies in the -cat mode of operation are done serially.